general: Initial commit
Initial commit of Jaktnät, an attempt to port the SerenityOS Jakt
language to .NET. This commit adds an ANTLRv4 grammar and enough
parser visitor code to successfully parse the Jakt "hello" sample.
An xUnit test project is set up as a test runner to run samples and
assert the output meets expectations, but this has not been fully
validated yet as the compiler is not yet emitting anything testable.
Right now, you can only debug via the unit test to confirm the parser
works.