compiler: Support (mutating) instance methods on classes.
This adds support for mutating instance (non-static) methods on classes.
Overloads are still not supported, but some refactoring is starting to
make that more possible. This change notably also changes Jakt classes
from being generated as .NET/C# Records to being regular C# classes
instead, as it is not possible to mutate a record's structural
properties.