parser/compiler: Add defer statement support
This uses a clever hack of an RAII-like Defer class I added to the
Runtime assembly that is IDisposable and takes a System.Action delegate
in its constructor. On Dispose, the Defer object invokes the Action.
This allows for passing in a lambda that closes over any local values.
By declaring it as a using statement variable declaration, the variable
is disposed when it goes out of scope, just like defer does in Jakt.

Note that there is an ugly hack of just naming the value with its
hashcode, which could have collisions and/or compiler errors if you try
to defer the same thing more than once.