compiler: First pass at inline C# support
This adds the basic inline C++ example from Jakt but is slightly
modified as inline C#. (It also removes the use of optionals which are
not yet implemented, and does not dereference the i32 as that is not
needed.

Currently, inline C# must be inside of an unsafe block, even though
unsafe blocks are not yet fully implemented. That is because you could
escape the Jaktnat safety mechanisms, not because you're using unsafe
code in C#. To use C# unsafe code, you would need to emit a C# unsafe
block within the inline C# code.